give receiver its own ffi file
